Rename generator.py to generation.py
