+++ tested/languages/java/config.py
@@ -20,7 +20,7 @@ from tested.languages.utils import jvm_cleanup_stacktrace, jvm_memory_limit, jvm
 from tested.serialisation import FunctionCall, Statement, Value
 
 if TYPE_CHECKING:
-    from tested.languages.generator import PreparedExecutionUnit
+    from tested.languages.generation import PreparedExecutionUnit
 
 
 logger = logging.getLogger(__name__)
