+++ tested/languages/python/config.py
@@ -17,7 +17,7 @@ from tested.languages.conventionalize import Conventionable, NamingConventions
 from tested.serialisation import FunctionCall, Statement, Value
 
 if TYPE_CHECKING:
-    from tested.languages.generator import PreparedExecutionUnit
+    from tested.languages.generation import PreparedExecutionUnit
 
 logger = logging.getLogger(__name__)
